% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccoef.r
\name{as.matrix.htlrfit}
\alias{as.matrix.htlrfit}
\title{Create a matrix of Markov chain samples}
\usage{
\method{as.matrix}{htlrfit}(x, k = x$K, ...)
}
\arguments{
\item{x}{An object of S3 class \code{htlrfit}.}

\item{k}{Coefficients associated with class \code{k} will be drawn. Must be a positive integer in 
1,2,\ldots,C for C-class traning labels. By default the last class will be chosen.}
}
\value{
A matrix with \code{(p + 1)} columns and \code{i} rows, where \code{p} is the number of features 
excluding intercept, and \code{i} is the number of iterations after burnin.
}
\description{
The Markov chain samples (without warmup) included in a \code{htlrfit} object will be coerced to a matrix.
}
\examples{
## No. of features used: 100; No. of iterations after burnin: 15 
fit <- htlr(X = colon$X, y = colon$y, fsel = 1:100, iter = 20, warmup = 5)

dim(as.matrix(fit))
  
}
