% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compred.r
\name{evaluate_pred}
\alias{evaluate_pred}
\title{Evaluate Prediction Results}
\usage{
evaluate_pred(y.pred, y.true, caseid = names(y.true), showplot = TRUE, ...)
}
\arguments{
\item{y.pred}{A matrix of predicted probabilities, as returned by a classifier.}

\item{y.true}{Ground truth labels vector.}

\item{caseid}{The names of test cases which we take account of. By default all test cases are used for evaluation.}

\item{showplot}{Logical; if \code{TRUE}, a summary plot will be generated.}

\item{...}{Not used.}
}
\value{
A summary of evaluation result.
}
\description{
This function compares the prediction results returned by a classifier with ground truth, 
and finally gives a summary of the evaluation.
}
