% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTLR-package.R
\docType{package}
\name{HTLR-package}
\alias{HTLR}
\alias{HTLR-package}
\title{Bayesian Logistic Regression with Heavy-Tailed Priors}
\description{
Efficient Bayesian multinomial logistic regression based on heavy-tailed priors. 
This package is suitable for classification and feature selection with high-dimensional 
features, such as gene expression profiles. Heavy-tailed priors can impose stronger 
shrinkage (compared to Gaussian and Laplace priors) to the coefficients associated with 
a large number of useless features, but still allow coefficients of a small number of 
useful features to stand out without punishment. It can also automatically make selection 
within a large number of correlated features. The posterior of coefficients and hyper-
parameters is sampled with restricted Gibbs sampling for leveraging high-dimensionality 
and Hamiltonian Monte Carlo for handling high-correlations among coefficients.
}
\references{
Longhai Li and Weixin Yao (2018). Fully Bayesian Logistic Regression 
with Hyper-Lasso Priors for High-dimensional Feature Selection.
\emph{Journal of Statistical Computation and Simulation} 2018, 88:14, 2827-2851.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://longhaisk.github.io/HTLR/}
  \item Report bugs at \url{https://github.com/longhaiSK/HTLR/issues}
}

}
\author{
\strong{Maintainer}: Longhai Li \email{longhai@math.usask.ca} (\href{https://orcid.org/0000-0002-3074-8584}{ORCID})

Authors:
\itemize{
  \item Steven Liu \email{xil865@usask.ca}
}

}
\keyword{internal}
