% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplrhmc.r
\name{htlr_fit}
\alias{htlr_fit}
\title{Fitting HTLR Models and Making Predictions}
\usage{
htlr_fit(y_tr, X_tr, X_ts = NULL, fsel = 1:ncol(X_tr), stdzx = TRUE,
  sigmab0 = 2000, ptype = "t", alpha = 1, s = -10, eta = 0,
  iters_h = 1000, iters_rmc = 1000, thin = 100, leap_L = 50,
  leap_L_h = 5, leap_step = 0.3, hmc_sgmcut = 0.05,
  initial_state = "lasso", alpha.rda = 0.2, silence = TRUE,
  .legacy = TRUE, predburn = NULL, predthin = 1)
}
\arguments{
\item{y_tr}{Vector of class labels in training or test data set. 
Must be coded as non-negative integers, e.g., 1,2,\ldots,C for C classes.}

\item{X_tr}{Design matrix of traning data; 
rows should be for the cases, and columns for different features.}

\item{fsel}{Subsets of features selected before fitting, such as by univariate screening.}

\item{stdzx}{If TRUE, the original features values are standardized to have mean = 0 and sd = 1.}

\item{initial_state}{The initial state of Markov Chain;
can be NULL, or a gived parameter vector, or a previous markov chain results.}
}
\value{
A list of fitting results.
}
\description{
This function trains linear logistic regression models with HMC in restricted Gibbs sampling. 
It also makes predictions for test cases if \code{X_ts} are provided.
}
\examples{
\dontrun{
data (gen_grpcor_data)
## creating training and test data sets
tr <- 200
X_tr <- data$X[1:tr,]
y_tr <- data$y[1:tr]
X_ts <- data$X[-(1:tr), ]
y_ts <- data$y[-(1:tr)]
## fit htlr models
fithtlr <- htlr_fitpred (
  y_tr = y_tr, X_tr = X_tr, stdzx = TRUE, #fsel = c (1,51,101),
  pty = "t", alpha = 1, s = -15, 
  iters_h = 1000, iters_rmc = 1000, thin = 50,
  leap_L_h = 5, leap_L = 50, leap_step = 0.5, hmc_sgmcut = 0.05, 
  initial_state = "bcbcsfrda", silence = F)
}

}
\references{
Longhai Li and Weixin Yao. (2018). Fully Bayesian Logistic Regression 
with Hyper-Lasso Priors for High-dimensional Feature Selection.
\emph{Journal of Statistical Computation and Simulation} 2018, 88:14, 2827-2851.
}
