% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccoef.r
\name{htlr_mccoef}
\alias{htlr_mccoef}
\title{Plots Markov chain trace or scatterplot}
\usage{
htlr_mccoef(fithtlr, features = 1, class = 2, usedmc = "all",
  symlim = FALSE, drawq = c(0, 1), truedeltas = NULL)
}
\arguments{
\item{fithtlr}{A list containing fitting results by \code{\link{htlr_fitpred}}.}

\item{features}{A vector of 1 or 2 numbers representing 1 or 2 features one wishes to look.}

\item{class}{Coefficients associated with \code{class} will be drawn. Must be a positive integer in 
1,2,\ldots,C for C-class traning labels.}

\item{usedmc}{Indices of Markov chain iterations used in plottings; one can set it to the 
indices of Markov chain iterations belonging to the ith feature subset, \code{mcids[[i]]}, 
found by \code{\link{htlr_fss}}.}
}
\value{
A vector of Markov chain sample of 1 coefficient, 
or an array of Markov chain samples of 2 coefficients.
}
\description{
This function plots Markov chain samples of 1 or 2 features. In plotting for 2 features, 
gray lines show Markov chain transitions.
}
