% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplrhmc.r
\name{htlr_predict}
\alias{htlr_predict}
\title{Make prediction on new data}
\usage{
htlr_predict(X_ts, fithtlr = NULL, deltas = NULL, burn = NULL,
  thin = NULL, usedmc = NULL)
}
\arguments{
\item{X_ts}{Matrix of values at which predictions are to be made.}

\item{fithtlr}{Fitted HTLR model object.}

\item{deltas}{The values of deltas (for example true deltas) used to prediction.}
}
\value{
A matrix of predictive probabilities, with rows for cases, cols for classes.
}
\description{
This function uses MCMC samples returned by \code{htlr_fit} to predict the class labels of test cases.
}
\seealso{
htlr_fit
}
