% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassorda.r
\name{lasso_deltas}
\alias{lasso_deltas}
\title{Lasso initial state}
\usage{
lasso_deltas(x, y, lambda = NA)
}
\arguments{
\item{y}{Vector of class labels in training or test data set. 
Must be coded as non-negative integers, e.g., 1,2,\ldots,C for C classes.}

\item{lambda}{A user supplied lambda value for \code{glmnet}. 
By default, it is determined by cross-validation.}

\item{X}{Design matrix of traning data; 
rows should be for the cases, and columns for different features.}
}
\value{
A matrix - the initial state of Markov Chain for HTLR model fitting.
}
\description{
Generate initial Markov chain state with Lasso.
}
\references{
Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010).
Regularization Paths for Generalized Linear Models via Coordinate
Descent. \emph{Journal of Statistical Software}, 33(1), 1-22.
}
\seealso{
htlr htlr_fit
}
\keyword{internal}
