% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.r
\name{order_ftest}
\alias{order_ftest}
\title{Order features by F-statistic}
\usage{
order_ftest(X, y)
}
\arguments{
\item{X}{Input matrix, of dimension \code{nobs} by \code{nvars}; each row is an observation vector.}

\item{y}{Vector of response variables.}
}
\value{
Order of all features of length \code{nvars}.
}
\description{
This function orders all features in terms of ANOVA F-statistic.
}
\examples{
data("diabetes392")
order_ftest(diabetes392$X, diabetes392$y)

}
\keyword{internal}
