% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.r
\name{order_kruskal}
\alias{order_kruskal}
\title{Order features by Kruskal-Wallis test}
\usage{
order_kruskal(X, y)
}
\arguments{
\item{X}{Input matrix, of dimension \code{nobs} by \code{nvars}; each row is an observation vector.}

\item{y}{Vector of response variables.}
}
\value{
Order of all features of length \code{nvars}.
}
\description{
This function orders all features in terms of Kruskal-Wallis test p-value.
}
\examples{
data("diabetes392")
order_kruskal(diabetes392$X, diabetes392$y)

}
\keyword{internal}
