% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.r
\name{order_plain}
\alias{order_plain}
\title{Plain order function}
\usage{
order_plain(X, y)
}
\arguments{
\item{X}{Input matrix, of dimension \code{nobs} by \code{nvars}; each row is an observation vector.}

\item{y}{Vector of response variables.}
}
\value{
Sequence starting from 1 to \code{nvars}.
}
\description{
A placeholder order function that returns the original order of given features.
}
\keyword{internal}
