% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{plot_fscore}
\alias{plot_fscore}
\title{Plots feature importance scores}
\usage{
plot_fscore(fscores, fsel = 1:length(fscores), show_ix = 0.1,
  do.plot = TRUE, ...)
}
\arguments{
\item{fscores}{Scores measuring feature importance, such as \code{wsdbs}, \code{msdbs}, or coefficients values.}
}
\description{
This function plots feature importance scores or coefficients using histogram line.
}
\seealso{
htlr_fss
}
