% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.htlr.fit}
\alias{predict.htlr.fit}
\title{Make Prediction on New Data}
\usage{
\method{predict}{htlr.fit}(object, newx, type = c("response", "class"), ...)
}
\arguments{
\item{object}{A fitted model object with S3 class \code{htlrfit}.}

\item{newx}{A Matrix of values at which predictions are to be made.}

\item{type}{Type of prediction required. Type "response" gives the fitted probabilities.
Type "class" produces the class label corresponding to the maximum probability.}

\item{...}{Advanced options to specify the Markov chain iterations used for inference. 
See \code{\link{htlr_predict}}.}
}
\value{
The object returned depends on type.
}
\description{
Similar to other predict methods, this function returns predictions from a fitted \code{htlrfit} object.
}
