% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{rank_ftest}
\alias{rank_ftest}
\title{Rank features by F-statistic}
\usage{
rank_ftest(X, y)
}
\arguments{
\item{X}{Input matrix, of dimension \code{nobs} by \code{nvars}; each row is an observation vector.}

\item{y}{Vector of response variables.}
}
\value{
Rank of all features of length \code{nvars}.
}
\description{
This function ranks all features in terms of ANOVA F-statistic.
}
\examples{
data("diabetes392")
rank_ftest(diabetes392$X, diabetes392$y)

}
