% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{rank_kruskal}
\alias{rank_kruskal}
\title{Rank features by Kruskal-Wallis test}
\usage{
rank_kruskal(X, y)
}
\arguments{
\item{X}{Input matrix, of dimension \code{nobs} by \code{nvars}; each row is an observation vector.}

\item{y}{Vector of response variables.}
}
\value{
Rank of all features of length \code{nvars}.
}
\description{
This function ranks all features in terms of Kruskal-Wallis test p-value.
}
\examples{
data("diabetes392")
rank_kruskal(diabetes392$X, diabetes392$y)

}
