% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{rank_plain}
\alias{rank_plain}
\title{Plain rank function}
\usage{
rank_plain(X, y)
}
\arguments{
\item{X}{Input matrix, of dimension \code{nobs} by \code{nvars}; each row is an observation vector.}

\item{y}{Vector of response variables.}
}
\value{
Sequence starting from 1 to \code{nvars}.
}
\description{
A placeholder rank function that returns the original order of given features.
}
\keyword{internal}
