% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{split_data}
\alias{split_data}
\title{Split data into train and test partitions}
\usage{
split_data(X, y, p.train = 0.7)
}
\arguments{
\item{X}{Input matrix, of dimension nobs by nvars; each row is an observation vector.}

\item{y}{Vector of response variables.}

\item{p.train}{Percentage of training set.}
}
\value{
List of training data \code{x.tr}, \code{y.tr} and testing data \code{x.te}, \code{y.te}.
}
\description{
This function splits the input data and response variables into training and testing parts.
}
\examples{
dat <- htlr_gendata_MLR(n = 100, p = 10)
dat <- split_data(dat$X, dat$y, p.train = 0.7)
dim(dat$x.tr)
dim(dat$x.te)
   
}
